# packages/contracts/openapi/central-rest.yaml
openapi: 3.1.0
info:
  title: Volleyball Central REST API
  version: 1.3.0
security:
  - sessionCookie: []
paths:
...
  /api/v1/analysis-runs/{analysisRunId}/overlay-manifest:
    get:
      summary: Get the authorized lazy browser overlay manifest
      parameters: [{ $ref: '#/components/parameters/AnalysisRunId' }]
      responses:
        '200':
          description: Versioned overlay manifest with bounded chunk URLs
          content: { application/json: { schema: { $ref: '../ai/overlay-manifest.schema.json' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/analysis-runs/{analysisRunId}/overlay-chunks/{chunkIndex}:
    get:
      summary: Get one authorized FlatBuffers browser overlay chunk
      parameters:
        - { $ref: '#/components/parameters/AnalysisRunId' }
        - { name: chunkIndex, in: path, required: true, schema: { type: integer, minimum: 0 } }
      responses:
        '200': { description: VOC1 overlay chunk, content: { application/vnd.volleyball.overlay-chunk+flatbuffers;version=1: {} } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    aiCallbackBearer:
      type: http
      scheme: bearer
  parameters:
...
    AnalysisRunId:
      name: analysisRunId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthenticated:
      description: Authentication required
      content: { application/json: { schema: { $ref: '../media/media-api-error.schema.json' } } }
    Forbidden:
      description: Resource is not accessible
      content: { application/json: { schema: { $ref: '../media/media-api-error.schema.json' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '../media/media-api-error.schema.json' } } }
...